Moving classes in separate packages
